+++ tfhe-benchmark/benches/integer/bench.rs
@@ -1386,15 +1386,13 @@ define_server_key_bench_default_fn!(
 mod cuda {
     use super::*;
     use benchmark::utilities::cuda_integer_utils::{cuda_local_keys, cuda_local_streams};
-    use criterion::{black_box, criterion_group};
+    use criterion::criterion_group;
     use std::cmp::max;
     use tfhe::core_crypto::gpu::{get_number_of_gpus, CudaStreams};
     use tfhe::integer::gpu::ciphertext::boolean_value::CudaBooleanBlock;
     use tfhe::integer::gpu::ciphertext::CudaUnsignedRadixCiphertext;
     use tfhe::integer::gpu::server_key::CudaServerKey;
     use tfhe::integer::{RadixCiphertext, ServerKey};
-    use tfhe::GpuIndex;
-    use tfhe_csprng::seeders::Seed;
 
     fn bench_cuda_server_key_unary_function_clean_inputs<F, G>(
         c: &mut Criterion,
@@ -1933,90 +1931,6 @@ mod cuda {
                 param,
                 param.name(),
                 "if_then_else",
-                &OperatorType::Atomic,
-                bit_size as u32,
-                vec![param.message_modulus().0.ilog2(); num_block],
-            );
-        }
-
-        bench_group.finish()
-    }
-
-    pub fn cuda_unsigned_oprf(c: &mut Criterion) {
-        let bench_name = "integer::cuda::unsigned_oprf";
-
-        let mut bench_group = c.benchmark_group(bench_name);
-        bench_group
-            .sample_size(15)
-            .measurement_time(std::time::Duration::from_secs(30));
-
-        for (param, num_block, bit_size) in ParamsAndNumBlocksIter::default() {
-            let param_name = param.name();
-
-            let bench_id;
-
-            match get_bench_type() {
-                BenchmarkType::Latency => {
-                    let streams = CudaStreams::new_multi_gpu();
-
-                    bench_id = format!("{bench_name}::{param_name}::{bit_size}_bits");
-                    bench_group.bench_function(&bench_id, |b| {
-                        let (cks, _cpu_sks) =
-                            KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-                        let gpu_sks = CudaServerKey::new(&cks, &streams);
-
-                        b.iter(|| {
-                            _ = black_box(
-                                gpu_sks
-                                    .par_generate_oblivious_pseudo_random_unsigned_integer_bounded(
-                                        Seed(0),
-                                        bit_size as u64,
-                                        num_block as u64,
-                                        &streams,
-                                    ),
-                            );
-                        })
-                    });
-                }
-                BenchmarkType::Throughput => {
-                    let (cks, cpu_sks) = KEY_CACHE.get_from_params(param, IntegerKeyKind::Radix);
-                    let gpu_sks_vec = cuda_local_keys(&cks);
-
-                    // Execute the operation once to know its cost.
-                    reset_pbs_count();
-                    cpu_sks.par_generate_oblivious_pseudo_random_unsigned_integer_bounded(
-                        Seed(0),
-                        bit_size as u64,
-                        num_block as u64,
-                    );
-                    let pbs_count = max(get_pbs_count(), 1); // Operation might not perform any PBS, so we take 1 as default
-
-                    bench_id = format!("{bench_name}::throughput::{param_name}::{bit_size}_bits");
-                    let elements = throughput_num_threads(num_block, pbs_count);
-                    bench_group.throughput(Throughput::Elements(elements));
-                    bench_group.bench_function(&bench_id, |b| {
-                        b.iter(|| {
-                            (0..elements).into_par_iter().for_each(|i| {
-                                let gpu_index: u32 = i as u32 % get_number_of_gpus();
-                                let stream = CudaStreams::new_single_gpu(GpuIndex::new(gpu_index));
-                                gpu_sks_vec[gpu_index as usize]
-                                    .par_generate_oblivious_pseudo_random_unsigned_integer_bounded(
-                                        Seed(0),
-                                        bit_size as u64,
-                                        num_block as u64,
-                                        &stream,
-                                    );
-                            })
-                        })
-                    });
-                }
-            }
-
-            write_to_json::<u64, _>(
-                &bench_id,
-                param,
-                param.name(),
-                "oprf",
                 &OperatorType::Atomic,
                 bit_size as u32,
                 vec![param.message_modulus().0.ilog2(); num_block],
@@ -2772,7 +2686,7 @@ mod cuda {
         cuda_trailing_zeros,
         cuda_trailing_ones,
         cuda_ilog2,
-        cuda_unsigned_oprf,
+        oprf::cuda::cuda_unsigned_oprf,
     );
 
     criterion_group!(
@@ -2800,7 +2714,7 @@ mod cuda {
         cuda_scalar_mul,
         cuda_scalar_div,
         cuda_scalar_rem,
-        cuda_unsigned_oprf,
+        oprf::cuda::cuda_unsigned_oprf,
     );
 
     criterion_group!(
